# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Taiyuan API",
    "version": "1.0.0",
    "description": "API Description"
  },
  "servers": [
    {
      "url": "https://api.taiyuan.dudustudio.monster"
    }
  ],
  "paths": {
...
    "/store-settings": {
      "get": {
        "tags": [
          "Store Settings"
        ],
        "summary": "讀取商店設置",
        "responses": {
          "200": {
            "description": "成功獲取商店設置",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "settings": {
                      "type": "object",
                      "additionalProperties": true
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "商店設置未找到"
          },
          "500": {
            "description": "伺服器錯誤"
          }
        }
      },
      "put": {
        "tags": [
          "Store Settings"
        ],
        "summary": "更新商店設置",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": true
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "商店設置更新成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "settings": {
                      "type": "object",
                      "additionalProperties": true
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "商店設置未找到"
          },
          "500": {
            "description": "伺服器錯誤"
          }
        }
      }
    },
    "/store-settings/logo/presigned-url": {
      "get": {
        "tags": [
          "Store Settings"
        ],
        "summary": "獲取商店 Logo 上傳用的預簽名 URL",
        "parameters": [
          {
            "name": "fileType",
            "in": "query",
            "required": true,
            "description": "文件類型 (.jpg, .png, .gif)",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功獲取預簽名 URL",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "uploadUrl": {
                      "type": "string",
                      "description": "用於上傳的預簽名 URL"
                    },
                    "imageUrl": {
                      "type": "string",
                      "description": "上傳後的圖片訪問 URL"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "不支持的文件類型"
          }
        }
      }
    },
    "/store-settings/logo": {
      "put": {
        "tags": [
          "Store Settings"
        ],
        "summary": "更新商店 Logo",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "imageUrl": {
                    "type": "string",
                    "description": "已上傳到 S3 的圖片 URL"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Logo 更新成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "logo": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "無效的圖片 URL"
          },
          "404": {
            "description": "商店設置未找到"
          }
        }
      }
    }
  },
...
}